requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
yt-dlp>=2024.1.0
groq>=0.4.2
python-dotenv==1.0.0
python-multipart==0.0.6
